+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Zero-order model
 ### Changed
 - `README.md` modified
+- Test system modified
 - Release range bug in `time_for_release` fixed
+- `matplotlib` import bug fixed
 ## [0.1] - 2025-07-27
 ### Added
 - Base model